+++ CHANGELOG.md
@@ -1,4 +1,12 @@
-## [0.1.0 - initial version
+## 0.1.4
+
+* add `FileCache.getString`
+
+## 0.1.3
+
+* add `FileCache.forceCacheSeconds` for force cached all url
+
+## 0.1.0
 
 * Dart2
 * remove httpClient